Moved everything for RegieLive into the providers file Match subtitle search even if a year is off by one (RL quirk)
